Put RIDF processors into own header
